Change the implementation of in-content movement text commands
Instead of javascript code, this new implementation mostly redirect
some keys press to others, e.g:

C-n: C-Left
C-n with mark: C-Shift-Left
...

So this way the browser does the actual movements. This allow the
bindings to work in contentEditable fields, like in gmail
textboxes. Also I hope this should help for #19 (text commands being
slow).

There is at least one minor drawback, if there is a selection in one
field and that the user move to another field (e.g. using the mouse)
the next movement commands will extends the selection (better behavior
would be that this would not be the case). But this is acceptable.