change default value of webview-container-stylesheet
The new style is made of black and red colors around the borders for
the active view, and should be more visible. It is only applied when
there is more than one view in the window. For non-active views, the
border width is applied so changing the focus does not redraw the web
buffer content.