keep QWebEngineView for a given buffer (start of refactoring)
QWebEngineViews are important as they store the context of the page, like
visible position in the page. Until now webmacs was moving QWebEnginePage from
view to view, but I think now it will be a lot better to keep views with page.

To refactor the code, the idea is that we keep the idea of WebView - but instead
of being a subclass of QWebEngineView, it is what the current code was calling
WebViewContainer. And for each webbuffer, the QWebEngineView is kept alive.