Use the qt network access manager to request webjump completions
This avoid using a complex code and handle the threaded stuff. One
downside is that the value given in webjump complete_fs must be
changed to take that in account (lists are still accepted, but they
will block the ui).

Another downside is that now some completion functions might block,
but this is an acceptable tradeoff I think to simplfy the code. Note
that overall the completion seems better at usage.