update adblock cache when the adblock-url-rules is changed
Fixes the github issue #41.

There was an issue with the current logic, for example if the list was reduced,
the cache was not rebuilt - the fix is to keep in a file the source urls that
were used to build the cache.

Note this does not handle the case if the urls ends with the same file name;
this should be fixed at some point too.