Fixes a crash when reviving closed buffer
This happened on some url, but not all - not sure why. Though the real issue is
that the code was loading the url and putting back the history stream. The fix
is to remove the load step, putting back the history stream is enough (and avoid
the segfault;)

I took the opportunity to refactor a bit the code around that feature.