Fix retrieving version in setup.py file
The version was hardcoded in the setup.py file, but this is wrong; Note also the
value was 1.0, but I never intended to set that version number...

A (hopefully not important) downside is that now the version will be less than
that (0.7). I hope this won't break anything out there.