Add revision information in webmacs://version
When available. At build time, a file "revision" is created and installed if the
installation is done from a git repository.

Then this file is used to extract the revision - if this file does not exists
but webmacs is directly in a git dir (e.g., from using pip install -e) then the
revision is extracted directly using the git command.