Fix using number keys in C-u f <number>
Calling commands is synchronous, and key prefix (C-u) was released after the
command was called - which prevented to use C-u or a number in prompts that were
started using the universal prefix.

The fix is to create the CommandContext a bit sooner, reset the prefix state,
then call the command.