Add where-is command bound to C-h w
This is meant to mimic the where-is command in emacs that lets you see on
what keys a certain command is bound.

Differently from emacs this will show where the command is bound not only
on currently active keymap(s), but it will show the command bindings in
all keymaps, annotated with the keymap name within parenthesis.