add support for passwordstore: the standard unix password manager
This supports reading passwords from the «pass» command line.

The variable «password-manager» must be set to "passwordstore" to enable it:
then the command "password-manager-fill-buffer" will try to fill the inputs of
the current web buffer using the pass information available.

The scheme of the password files must follow the one defined in
https://github.com/passff/passff. To resume,

mypassword
login: username
url: part-of-the-url-to-connect

Other fields are also supported as described in passff documentation.

Notes:

- this is not enabled by default.
- for backward compatibility, the autofill database (custom way to save/autofill
  passwords using a db file in the session directory) is still active. Though
  the API and both the use case are quite similar, and should be unified at some
  point.
- no public documentation yet