Remove autofill code;
Not sure how it was used in the wild, and not sure it was working correctly
anyway.

This is a feature (or set of features??) I don't use, and unless someone
complains hard I am willing to completely remove the code.

Personnaly, I use passwordstore instead, and I'd encourage people to do so.